refact(ops,project): move list_roots to project
Every thing related to roots should be in that module.
